refactor: 패키지 의존성 관리를 위해 DTO 위치를 domain 패키지로 옮김.
- 기존처럼 web 패키지에 DTO 가 위치하면 service 가 web 패키지에 의존하게 되고, 이는 controller -> service -> repository 의존 흐름에 맞지 않음.
