+++ src/main/java/com/programmers/jpaboard/domain/post/dto/PostUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.programmers.jpaboard.web.post.dto;
+package com.programmers.jpaboard.domain.post.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
